+++ src/main/java/ptit/d19cqcp02/webMVC/model/dto/OrderDetailDTO.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.DecimalMin;
 @Getter
 @Setter
 public class OrderDetailDTO {
-    private Long orderId;
+    private Integer orderId;
     private Long productId;
     @DecimalMin(value = "1")
     private Integer amount;
