+++ src/main/java/ptit/d19cqcp02/webMVC/model/entity/OrderDetailView.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderDetailView {
-    Long orderId;
+    Integer orderId;
 
     String firstName;
